--warnings ++tests/cpp compile
the cpp tests fails because of a bug with <ENTER> on line 0 for now ...

+++ action.cpp
@@ -401,11 +401,11 @@ YZCursor YZAction::search( YZView* pView, const QString& what, const YZCursor& m
 
 		if ( idx >= 0 ) {
 			//check we are not past the mEnd
-			if ( i == mEnd.getY() && ((!reverseSearch && idx > mEnd.getX()) || (reverseSearch && idx < mEnd.getX()))) 
+			if ( i == mEnd.getY() && ((!reverseSearch && (unsigned int)idx > mEnd.getX()) || (reverseSearch && (unsigned int)idx < mEnd.getX()))) 
 				continue; //should exit the loop at next iteration anyway and return false
 
 			//i really found it ? or is it a previous "found" ?
-			if ( mBegin.getX() == ( unsigned int ) idx && reverseSearch && i == currentMatchLine ) { //ok we did not move guy (col 0 or last col maybe ...)
+			if ( mBegin.getX() == ( unsigned int ) idx && reverseSearch && i == (unsigned int)currentMatchLine ) { //ok we did not move guy (col 0 or last col maybe ...)
 				yzDebug() << "Only a fake match on this line, skip it" << endl;
 				if ( reverseSearch )
 					currentMatchColumn=-1;
